v1.0.9
- Reduced webservers to 1. All webservice calls and static files hosting use the same webserver now. Not sure why I did this to begin with but it was inefficient.
- Readjusted all code to use dynamic port numbers. No longer hard code port 3000 for web service calls.
-  Added "BambuBoard_logging" config setting for server side logging switch.
