+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "bambuboard",
-  "version": "1.0.8",
+  "version": "1.0.9",
   "description": "bambu x1 printer dashboard",
   "main": "bambuConnection.js",
   "keywords": [],
