+++ public/widgets/bed-temp/index.html
