+++ public/widgets/fans/index.html
