+++ public/widgets/model-image/model-script.js
@@ -2,6 +2,7 @@
 /// Configure your settings here:
 
 const serverURL = window.location.hostname; // IP of the computer running this dashboard
+const serverPort = window.location.port;
 
 // Note: If set to 127.0.0.1 you will not be able to view your plate image, weight or total prints.
 //       Those features will only work if viewing the dashboard locally.
@@ -109,7 +110,7 @@ executeTask();
   // Send credentials to your own server
   async function loginAndFetchImage() {
     try {
-        const response =  await fetch('http://' + serverURL + ':3000/login-and-fetch-image', {
+        const response =  await fetch('http://' + serverURL + ':' + serverPort + '/login-and-fetch-image', {
             method: 'GET',
             headers: { 'Content-Type': 'application/json' }
         });
