+++ public/widgets/notes/edit.html
@@ -19,7 +19,7 @@
        async function loadNote() {
             try {
                 const serverURL = window.location.hostname;
-                const response = await fetch('http://' + serverURL + ':3000/note');
+                const response = await fetch('http://' + serverURL + ':' + serverPort + '/note');
                 if (response.ok) {
                     const data = await response.json();
                     document.getElementById('noteContent').innerText = data.content;
@@ -37,7 +37,7 @@
             const content = document.getElementById('noteContent').value;
             const serverURL = window.location.hostname;
             try {
-                const response = await fetch('http://' + serverURL + ':3000/note', {
+                const response = await fetch('http://' + serverURL + ':' + serverPort + '/note', {
                     method: 'PUT',
                     headers: {
                         'Content-Type': 'application/json',
