+++ public/widgets/notes/index.html
@@ -16,7 +16,7 @@
         async function loadNote() {
             try {
                 const serverURL = window.location.hostname;
-                const response = await fetch('http://' + serverURL + ':3000/note');
+                const response = await fetch('http://' + serverURL + ':' + serverPort + '/note');
                 if (response.ok) {
                     const data = await response.json();
                     document.getElementById('noteDisplay').innerText = data.content;
