+++ public/widgets/nozzle-info/index.html
