+++ public/widgets/nozzle-temp/index.html
