+++ public/widgets/progress-info/index.html
