+++ public/widgets/wifi/index.html
